test(pgtap): fix 4 individual test bugs (A5)
- updated_at.sql: add seed_period_criteria() so score_sheet_items
  INSERT gets a non-NULL period_criterion_id (LIMIT 1 returned NULL)
- admin_clone_framework.sql: seed source framework a1b2c3d4-e5f6-...
  before clone tests; fix plan(8)→plan(7) (file has 7 assertions)
- admin_generate_entry_token.sql: cast COUNT(*)::int to fix
  is(bigint, integer) overload ambiguity
- org_admin_cancel_invite.sql: change ON CONFLICT (id) to
  ON CONFLICT (user_id, organization_id) to handle admin A already
  having an active membership in org A from seed_two_orgs()

Refs: docs/superpowers/plans/ci-e2e-remediation/phase-a-pgtap-fixes.md

+++ sql/tests/rpcs/contracts/admin_generate_entry_token.sql
@@ -104,7 +104,7 @@ SELECT rpc_admin_generate_entry_token('cccc0000-0000-4000-8000-000000000011'::uu
 
 -- Count non-revoked tokens for this period; should be exactly 1
 SELECT is(
-  (SELECT COUNT(*) FROM entry_tokens
+  (SELECT COUNT(*)::int FROM entry_tokens
    WHERE period_id = 'cccc0000-0000-4000-8000-000000000011'::uuid
      AND is_revoked = false),
   1,
